클라이언트 - 서버 간 채팅 패킷 처리, 채팅 개행 개선

+++ Server/Creature.h
@@ -14,7 +14,7 @@ public:
 
 	void OnDamaged(CreatureRef attacker);
 	void KnockBack();
-	uint64 SetWait(float time) { return _waitHit = GetTickCount64() + time; }
+	uint64 SetWait(uint64 time) { return _waitHit = GetTickCount64() + time; }
 
 protected:
 	uint64 _waitUntil = 0;
